+++ packages/components/src/overlay/Dropdown/Item/Item.constants.js
@@ -1,10 +1,13 @@
 import PropTypes from 'prop-types';
 
-export const ITEM_DEFAULT_PROPS = {};
+export const ITEM_DEFAULT_PROPS = {
+  disabled: false,
+};
 export const ITEM_PROP_TYPES = {
   label: PropTypes.string,
   image: PropTypes.string,
   icon: PropTypes.node,
   dataSelected: PropTypes.bool,
   group: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+  disabled: PropTypes.bool,
 };