+++ packages/leemons/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @bubbles-ui/leemons
 
+## 1.2.117
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.117
+  - @bubbles-ui/calendars@1.2.117
+  - @bubbles-ui/editors@1.2.117
+  - @bubbles-ui/icons@1.2.117
+
+## 1.2.116
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.116
+  - @bubbles-ui/calendars@1.2.116
+  - @bubbles-ui/editors@1.2.116
+  - @bubbles-ui/icons@1.2.116
+
 ## 1.2.115
 
 ### Patch Changes